+++ src/main/java/com/jm/jobseekerplatform/dao/impl/JobExperienceDAO.java
@@ -12,13 +12,11 @@ import java.util.Set;
 @Repository("jobExperienceDAO")
 public class JobExperienceDAO extends AbstractDAO<JobExperience> {
 
-    public Set<Long> getAllExperiencesIdForResume(Long id){
+    public Set<BigInteger> getAllExperiencesIdForResume(Long id){
         Query query = entityManager.createNativeQuery("select job_experiences_id from resumes_job_experiences where " +
                 "resume_id = :id").setParameter("id", id);
-        Set<Long> experiencesId = new HashSet<>();
-        query.getResultList().forEach(element ->
-                experiencesId.add(((BigInteger)element).longValue())
-        );
+        Set<BigInteger> experiencesId = new HashSet<>();
+        query.getResultList().forEach(element -> experiencesId.add((BigInteger)element));
         return experiencesId;
     }
 }
